% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonify.R
\name{chart2music}
\alias{chart2music}
\alias{use_chart2music}
\alias{html_dependency_chart2music}
\title{chart2music}
\usage{
use_chart2music()

html_dependency_chart2music()
}
\value{
An \code{htmltools::tagList()} with the chart2music dependencies, or an
\link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
Slide tone plays a subtle sound when you change slides.The tones increase in
pitch for each slide from a low C to a high C note. The tone pitch stays the
same for incremental slides.
}
\section{Functions}{
\itemize{
\item \code{use_chart2music()}: Adds sonification to your ggplot.

\item \code{html_dependency_chart2music()}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the
chart2music dependencies. Most users will want to use \code{use_chart2music()}.

}}
\section{Usage}{
 To add chart2music to your ggplot, add the
following code chunk to your R Markdown file.

\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r data11y-chart2music, echo=FALSE\}
data11y::use_chart2music()
```
}\if{html}{\out{</div>}}
}

\examples{
use_chart2music()

}
\references{
\href{https://chart2music.com}{chart2music}
}
