% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{create_barplot_schema}
\alias{create_barplot_schema}
\title{Create a barplot schema}
\usage{
create_barplot_schema(
  type1,
  id,
  title,
  xlabel,
  ylabel,
  xticklabels,
  yticklabels,
  ymax
)
}
\arguments{
\item{type1}{A character string indicating the type of the plot.}

\item{id}{An integer indicating the ID of the plot.}

\item{title}{A character string indicating the title of the plot.}

\item{xlabel}{A character string indicating the label for the x-axis.}

\item{ylabel}{A character string indicating the label for the y-axis.}

\item{xticklabels}{A character vector containing the labels for the x-axis ticks.}

\item{yticklabels}{A character vector containing the labels for the y-axis ticks.}

\item{ymax}{A numeric value indicating the maximum value for the y-axis.}
}
\value{
A json object containing the schema for the barplot.
}
\description{
This function creates a schema for a barplot with customizable parameters.
}
