% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{create_scatterplot_schema}
\alias{create_scatterplot_schema}
\title{Create a scatterplot schema}
\usage{
create_scatterplot_schema(id, title, xlabel, ylabel, layers)
}
\arguments{
\item{id}{An integer indicating the ID of the plot.}

\item{title}{A character string indicating the title of the plot.}

\item{xlabel}{A character string indicating the label for the x-axis.}

\item{ylabel}{A character string indicating the label for the y-axis.}

\item{type1}{A character string indicating the type of the plot.}

\item{xticklabels}{A character vector containing the labels for the x-axis ticks.}

\item{yticklabels}{A character vector containing the labels for the y-axis ticks.}

\item{ymax}{A numeric value indicating the maximum value for the y-axis.}
}
\value{
A json object containing the schema for the scatterplot.
}
\description{
This function creates a schema for a scatterplot with customizable parameters.
}
