% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maidr-package.R
\docType{package}
\name{maidr-package}
\alias{maidr}
\alias{maidr-package}
\title{maidr: Multimodal Access and Interactive Data Representation}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/data11y}
  \item Report bugs at \url{https://github.com/jooyoungseo/data11y/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jseo1005@illinois.edu} (\href{https://orcid.org/0000-0002-4064-6012}{ORCID})

}
\keyword{internal}
