% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_svg}
\alias{print_svg}
\title{Print an SVG file}
\usage{
print_svg(g, id = "plot", width = 8, height = 3.75, ...)
}
\arguments{
\item{g}{A plot2 object.}

\item{id}{A character string specifying the ID of the SVG element.}

\item{width}{A numeric value specifying the width of the SVG element in inches.}

\item{height}{A numeric value specifying the height of the SVG element in inches.}

\item{...}{Additional arguments to be passed to the \code{\link[grDevices]{svg}} function.}
}
\value{
SVG object
}
\description{
This function prints an SVG from a plot object.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
print_svg(p, id = "myplot", width = 10, height = 5, browse = TRUE)

}
